Add Static Analyzer Annotations
This adds support files and documentaiton for the missing-GC-root
finding analysis passes at https://github.com/Keno/ClangSA.jl.
In order to be effective, the codebase needs to be annotated with
various annotations, which this commit introduces and documents.
However, since there are many annotations required, there would
be many conflicts if done in one PR, so I'm doing these incrementally.